# learning/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Resource Scheduling Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/availability/day": {
      "get": {
        "tags": [
          "availability"
        ],
        "summary": "Day Availability",
        "description": "Free/busy status of every active resource on `date`.\n\nPowers the room-list colour dots (`is_free`) and the empty-room search filter.",
        "operationId": "day_availability_api_v1_availability_day_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Calendar day to check (YYYY-MM-DD)",
              "title": "Date"
            },
            "description": "Calendar day to check (YYYY-MM-DD)"
          },
          {
            "name": "resource_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ResourceType"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional filter",
              "title": "Resource Type"
            },
            "description": "Optional filter"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResourceAvailability"
                  },
                  "title": "Response Day Availability Api V1 Availability Day Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/availability/free-slots": {
      "get": {
        "tags": [
          "availability"
        ],
        "summary": "Free Slots",
        "description": "Open windows on one resource for `date`, at least `duration_minutes` long,\nsearched inside the working window [from_hour, to_hour).",
        "operationId": "free_slots_api_v1_availability_free_slots_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "resource_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Day to search (YYYY-MM-DD)",
              "title": "Date"
            },
            "description": "Day to search (YYYY-MM-DD)"
          },
          {
            "name": "duration_minutes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1440,
              "minimum": 15,
              "default": 60,
              "title": "Duration Minutes"
            }
          },
          {
            "name": "from_hour",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 23,
              "minimum": 0,
              "description": "Working-window start hour (UTC)",
              "default": 8,
              "title": "From Hour"
            },
            "description": "Working-window start hour (UTC)"
          },
          {
            "name": "to_hour",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 24,
              "minimum": 1,
              "description": "Working-window end hour (UTC)",
              "default": 20,
              "title": "To Hour"
            },
            "description": "Working-window end hour (UTC)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FreeSlotsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FreeSlotsOut": {
        "properties": {
          "resource_id": {
            "type": "string",
            "title": "Resource Id"
          },
          "free_slots": {
            "items": {
              "$ref": "#/components/schemas/Interval"
            },
            "type": "array",
            "title": "Free Slots"
          }
        },
        "type": "object",
        "required": [
          "resource_id",
          "free_slots"
        ],
        "title": "FreeSlotsOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Interval": {
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "title": "End"
          }
        },
        "type": "object",
        "required": [
          "start",
          "end"
        ],
        "title": "Interval"
      },
...
      "ResourceAvailability": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "resource_type": {
            "$ref": "#/components/schemas/ResourceType"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "capacity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity"
          },
          "requires_approval": {
            "type": "boolean",
            "title": "Requires Approval"
          },
          "is_free": {
            "type": "boolean",
            "title": "Is Free"
          },
          "busy": {
            "items": {
              "$ref": "#/components/schemas/Interval"
            },
            "type": "array",
            "title": "Busy"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "resource_type",
          "requires_approval",
          "is_free",
          "busy"
        ],
        "title": "ResourceAvailability"
      },
...
      "ResourceType": {
        "type": "string",
        "enum": [
          "classroom",
          "lab",
          "computer_room",
          "seminar_hall",
          "meeting_room",
          "equipment",
          "other"
        ],
        "title": "ResourceType",
        "description": "Why enum and not a free-form string?\nControlled vocabulary prevents drift (\"Lab\" vs \"lab\" vs \"Laboratory\").\nNew types added here \u2014 never hardcoded in business logic."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}